+++ example/callContract.dart
@@ -8,7 +8,7 @@
 
 // void main() async {
 //   // init web3 module
-//   Web3 web3 = new Web3(
+//   Web3 web3 = Web3(
 //     approveCb: approvalCallback,
 //     networkId: 122,
 //     url: 'https://rpc.fuse.io',
@@ -26,7 +26,7 @@
 //   print('account address: $accountAddress');
 
 //   // init api module
-//   API api = new API('https://studio.fuse.io/api');
+//   API api = API('https://studio.fuse.io/api');
 
 //   api.setJwtToken('YOUR_JWT');