+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-# work in progress
+# v0.4.0
+
+## New Feature
+- Rename API module to StudioApi & added api key
+- `api.setJwtToken` is deprecated